gateway-requirements.txt
flask==2.3.3
flask-cors==4.0.0
grpcio==1.60.1
grpcio-tools==1.60.1